# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TV Carlos API
  version: 3.0.0
  description: API pública e administrativa da Web TV. Escritas exigem cookie de sessão e token CSRF.
servers:
  - url: http://localhost:8082
    description: Docker local
...
security: []
paths:
...
  /api/auth/session:
    get:
      tags: [Auth]
      operationId: getSession
      security: [{ sessionCookie: [] }]
      responses:
        '200': { description: Sessão atual, content: { application/json: { schema: { $ref: '#/components/schemas/Session' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/upload/image:
    post:
      tags: [Media]
      operationId: uploadImage
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody: { $ref: '#/components/requestBodies/File' }
      responses: { '201': { description: Imagem validada e armazenada, content: { application/json: { schema: { $ref: '#/components/schemas/StoredMedia' } } } }, '400': { $ref: '#/components/responses/ValidationError' }, '413': { description: Limite de 10 MB excedido } }
  /api/upload/video:
    post:
      tags: [Media]
      operationId: uploadVideo
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody: { $ref: '#/components/requestBodies/File' }
      responses: { '201': { description: Vídeo validado por assinatura e ffprobe, content: { application/json: { schema: { $ref: '#/components/schemas/StoredMedia' } } } }, '400': { $ref: '#/components/responses/ValidationError' }, '413': { description: Limite de 500 MB excedido } }
  /api/upload/document:
    post:
      tags: [Media]
      operationId: uploadDocument
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody: { $ref: '#/components/requestBodies/File' }
      responses: { '201': { description: PDF validado por assinatura e armazenado para material de apoio, content: { application/json: { schema: { $ref: '#/components/schemas/StoredMedia' } } } }, '400': { $ref: '#/components/responses/ValidationError' }, '413': { description: Limite de 25 MB excedido } }
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: tv_session }
    privateRoomCookie: { type: apiKey, in: cookie, name: tv_private_room_session }
    csrfHeader: { type: apiKey, in: header, name: X-CSRF-Token }
...
  requestBodies:
...
    File: { required: true, content: { multipart/form-data: { schema: { type: object, required: [file], properties: { file: { type: string, format: binary } } } } } }
  responses:
    ValidationError: { description: Entrada rejeitada, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unauthorized: { description: Sessão ausente ou expirada, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
            type: object
            required: [code, message]
            properties: { code: { type: string }, message: { type: string } }
...
    Session: { type: object, required: [user, expiresAt], properties: { user: { type: object, required: [id, username, role], properties: { id: { type: string, format: uuid }, username: { type: string }, role: { type: string, enum: [admin, teacher] } } }, expiresAt: { type: string, format: date-time } } }
...
    StoredMedia: { type: object, required: [key, publicUrl], properties: { key: { type: string }, publicUrl: { type: string }, filename: { type: string } } }